requirements.txt
gradio
huggingface_hub[hf_xet]
smolagents
smolagents[litellm]
smolagents[openai]
langchain
langchain_community
langchain-ollama
langgraph
langchain-anthropic
langchain-litellm
langgraph-swarm
langgraph-supervisor
ollama
litellm
arxiv
wikipedia-api
duckduckgo-search
chromadb
pymupdf
pandas
bs4
markdownify
openpyxl
matplotlib
moviepy
numpy
pillow
pydub
python-dotenv
requests
scikit-learn
sentence-transformers
seaborn
SpeechRecognition
tqdm
transformers
pypdf